requirements.txt
pillow
opencv-python
matplotlib
numpy
pyserial